add stub for inject_embed_test
