+++ test/experimental/inject_embed_test.html
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+
+<!-- Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
+     for details. All rights reserved. Use of this source code is governed by a
+     BSD-style license that can be found in the LICENSE file. -->
+
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <title>Embeddings Demo</title>
+
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap-theme.min.css">
+
+    <link href='https://fonts.googleapis.com/css?family=Inconsolata:400,700'
+          rel='stylesheet' type='text/css'>
+
+    <!-- Test package stuff -->
+    <link rel="x-dart-test" href="inject_embed_test.dart">
+    <script src="packages/test/dart.js"></script>
+    <style>
+        .dartpad-embed > * {
+            width: 100%;
+        }
+
+        iframe {
+            width: 100%;
+            height: 400px;
+            margin-bottom: 40px;
+            margin-top: 40px;
+        }
+    </style>
+</head>
+
+<body>
+<div class="container">
+    <div class="row">
+        <div class="col-md-10">
+            <h4>Hello, World!</h4>
+            <p>
+                Below is a 'Hello, World!' app.
+            </p>
+            <div class="dartpad-embed">
+                <pre>
+                    <code>foo</code>
+                </pre>
+            </div>
+        </div>
+    </div>
+</div>
+</body>
+</html>
